lab2/lab2/backend/requirements.txt
Django
djangorestframework
markdown
django-filter
django-cors-headers
psycopg[binary,pool]
faker
gunicorn